DigitalGov Logo Replaement in Header
Per email from Sarah:  "The official branding on the Midas platform should read Open Opportunities for DigitalGov.

When the new platform is launched we will be adding a box or banner on Digitalgov that links to Open Opportunities for Digitalgov. The box or banner will be updated regularly highlighting new tasks, notes on participants and other info to pique interest."